Added animals to search
Get both events and animals in search results. Added touchable opacity to the results to link to their respective tabs.  Clicking on an event only takes you to the hope page because events doesn't have its own tab. Fixed both errors related to the previous implementation of the search feature.